Remove the hard depdendency on Redis
Use of Redis with the ButterCMS gem is optional, so a hard dependency doesn't make sense. It's not fun to get Redis bundled into your project if you're not using it.

This PR removes the hard dependency on Redis and moves the data store requires to where they are invoked, so the optional dependency on Redis can be enforced only if the user actually makes use of it.