Switch to Net::HTTP from HTTParty
Gems should strive to use the built in Net::HTTP library instead of
bringing in a 3rd party library. Net::HTTP is pretty great these days
and the API is simple enough once you've worked with it for a while.

Given that ButterCMS is not taking advantage of the abstractions that
HTTParty provides it seems like overkill to have to bring in yet another
dependency into my application.